Create homework3.md
Mike Johnson spent $228,665 in total contributions.  That includes individual contributions of itemized and unitemized individual contributions. Johnson is running for Louisiana Representative position.

In the total disbursements category, this committee has spent $74, 039.77 in the span of seven months starting January 1, 2017. In the documents that had to be turned in, this amount included lodging, sponsorship, website, fundraising services, etc.  What can be a shocking is that Johnson sent over $4,000 in contribution to Rennaci for Ohio to support. 

Johnson's campaign did not shy away from spending big money when it came to fundraisers either. $1350 went to the housing fund. Almost $3600 went to a campaign consulting firm. 

All in all, there was nothing out of the ordinary when it comes to his campaign finances. He did a lot of his traveling within the south, specifically Louisiana. In that case, it is not expensive for lodging at all.