Add files via upload
This is the chart where I'm explaining my journey to get back  in shape by running and hiking. As you see, the more miles I ran, the less I hiked. I would never do both in the same day, of course. 2016 was a big fitness year. Now in 2017, I'm preparing for a marathon in November.