Create homework4.md
/Users/astasiawilliams/Desktop/Running_Hiking_chartbuilder.png

This chart shows my tracking of how my running and hiking training for my first marathon earlier this year. As you see the more mileage I ran, the less miles I did while hiking. 2016 was a big year of trying to get back healthy. Now in 2017 I am going to train for a marathon In January 2018. ;